Separating result processor out from profiler.py without changing the behaviors of current profile.py
